Split extensions into separate files and moved to extensions folder
