+++ src/Api/MappingProfile.cs
@@ -8,15 +8,26 @@ namespace AuthorizationManagement.Api
         public MappingProfile()
         {
             CreateMap<Models.Application, Application>()
+                .ForMember(g => g.ETag, cfg => cfg.Ignore())
+                .ForMember(g => g.LastModifiedOn, cfg => cfg.Ignore())
                 .ReverseMap();
 
             CreateMap<Models.Group, Group>()
+                .ForMember(g => g.ApplicationId, cfg => cfg.Ignore())
+                .ForMember(g => g.ETag, cfg => cfg.Ignore())
+                .ForMember(g => g.LastModifiedOn, cfg => cfg.Ignore())
                 .ReverseMap();
 
             CreateMap<Models.User, User>()
+                .ForMember(u => u.ApplicationId, cfg => cfg.Ignore())
+                .ForMember(u => u.ETag, cfg => cfg.Ignore())
+                .ForMember(u => u.LastModifiedOn, cfg => cfg.Ignore())
                 .ReverseMap();
 
             CreateMap<Models.UserGroup, UserGroup>()
+                .ForMember(ug => ug.ApplicationId, cfg => cfg.Ignore())
+                .ForMember(ug => ug.ETag, cfg => cfg.Ignore())
+                .ForMember(ug => ug.LastModifiedOn, cfg => cfg.Ignore())
                 .ReverseMap();
         }
     }
